Refactor into a single process
