+++ spy_relay/package.json
@@ -6,9 +6,7 @@
   "scripts": {
     "build": "tsc",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "spy_relay": "node lib/spy_relay.js",
-    "spy_rest": "node lib/spy_rest.js",
-    "spy_worker": "node lib/spy_worker.js"
+    "spy_relay": "node lib/spy_relay.js"
   },
   "author": "",
   "license": "Apache-2.0",
